Add preliminary RSKextractprofiles and RSKplotprofiles.
extractprofiles is probably temporary -- a better route might be to
just read the profile events during RSKopen, so that a function
equivalent to RSKreaddata (i.e. RSKreadprofiles) could be used to read
*only* the profiles, and in particular only the *desired* profiles.

Perhaps RSKreadprofiles and RSKreaddata should be mutually exclusively
operating commands, which populate different structures, so one
*could* use both of them to read the "time series" data as well as the
profiles, but there should be a caveat emptor that doing so will use
significantly more memory, which might be an issue for large datasets.